% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ac_A1R12_1}
\alias{ac_A1R12_1}
\title{ac_A1R12_1}
\format{
A tibble with the columns:
\describe{
\item{MO}{Microorganism},
\item{met}{Metabolite}
\item{flux}{Flux of the metabolite}}
}
\source{
\url{https://github.com/sirno/misosoup}
}
\usage{
ac_A1R12_1
}
\description{
Example community taken from \code{misosoup}.
}
\examples{
ac_A1R12_1

}
\keyword{datasets}
