% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsmc-data.R
\docType{data}
\name{ac_A1R12_2}
\alias{ac_A1R12_2}
\title{ac_A1R12_2}
\format{
A tibble with the columns:
\describe{
\item{species}{Microorganisms},
\item{metabolites}{Metabolites}
\item{fluxes}{Fluxes of the metabolite}}
}
\source{
\url{https://github.com/sirno/misosoup}
}
\usage{
ac_A1R12_2
}
\description{
Example community taken from \code{misosoup}.
}
\examples{
ac_A1R12_2

}
\keyword{datasets}
