% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ac_A3R04_1}
\alias{ac_A3R04_1}
\title{Example Data from MiSoSoup}
\format{
A tibble with the columns:
\describe{
\item{species}{Microorganisms},
\item{metabolites}{Metabolites}
\item{fluxes}{Fluxes of the metabolite}
}
}
\usage{
ac_A3R04_1
}
\description{
A3R04 focal strain in acetate medium.
}
\keyword{datasets}
