% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-align.R
\name{align}
\alias{align}
\alias{align,MiCoAl-method}
\title{Align Microbial Community Objects}
\usage{
align(object, pairwise = TRUE)

\S4method{align}{MiCoAl}(object, pairwise = TRUE)
}
\arguments{
\item{object}{A list (?) of objects of type MiCo}

\item{pairwise}{Logical to toggle pairwise alignment}
}
\value{
An overview over the alignment of two microbial communities.
}
\description{
Align Microbial Community Objects
}
\section{Functions}{
\itemize{
\item \code{align(MiCoAl)}: Align Microbial Community Objects

}}
\examples{
### Todo
}
