% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-alignmentNetwork.R
\name{alignmentNetwork}
\alias{alignmentNetwork}
\title{MiCoAl Network Plot}
\usage{
alignmentNetwork(object, frac)
}
\arguments{
\item{object}{An object of class MiCoAl.}

\item{frac}{Minimum relative weight of maximum weight to plot.}
}
\value{
An igraph network
}
\description{
MiCoAl Network Plot
}
\examples{
###
}
