% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsmc-data.R
\docType{data}
\name{cit_A1R12_1}
\alias{cit_A1R12_1}
\title{cit_A1R12_1}
\format{
A tibble with the columns:
\describe{
\item{species}{Microorganisms},
\item{metabolites}{Metabolites}
\item{fluxes}{Fluxes of the metabolite}}
}
\source{
\url{https://github.com/sirno/misosoup}
}
\usage{
cit_A1R12_1
}
\description{
Example community taken from \code{misosoup}.
}
\examples{
cit_A1R12_1

}
\keyword{datasets}
