% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cit_A1R12_2}
\alias{cit_A1R12_2}
\title{Example Data from MiSoSoup}
\format{
A tibble with the columns:
\describe{
\item{species}{Microorganisms},
\item{metabolites}{Metabolites}
\item{fluxes}{Fluxes of the metabolite}
}
}
\usage{
cit_A1R12_2
}
\description{
A1R12 focal strain in citrate medium.
}
\keyword{datasets}
