% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareAlignments.R
\name{compareAlignments}
\alias{compareAlignments}
\title{Compare Alignments}
\usage{
compareAlignments(
  ...,
  names,
  smooth = FALSE,
  se = FALSE,
  min_frac = NULL,
  max_frac = NULL
)
}
\arguments{
\item{...}{MiCoAl objects}

\item{names}{Character vector supplying names to be used as labels in plot}

\item{smooth}{Boolean to toggle a smooth line}

\item{se}{Boolean to toggle std. error bands}

\item{min_frac}{Numerical value specifying minimum fraction in alignment}

\item{max_frac}{Numerical value specifying maximum fraction in alignment}
}
\value{
A ggplot for the comparison of multiple MiCoAl objects.
}
\description{
Visual comparison of multiple alignments by plotting the number of aligned
reactions against the fraction of aligned communities
}
