% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignMultiple.R
\name{.alignMultiple}
\alias{.alignMultiple}
\title{Multiple Alignment of MiCo (Microbial Community) Objects}
\usage{
.alignMultiple(coms)
}
\arguments{
\item{coms}{Named list of MiCo objects.}
}
\value{
A list containing the alignment data.
}
\description{
Multiple Alignment of MiCo (Microbial Community) Objects
}
