% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignPairwise.R
\name{.alignPairwise}
\alias{.alignPairwise}
\title{Pairwise Alignment of MiCo (Microbial Community) Objects}
\usage{
.alignPairwise(coms)
}
\arguments{
\item{coms}{Named list of MiCo objects.}
}
\value{
List of pairwise alignments.
}
\description{
Pairwise Alignment of MiCo (Microbial Community) Objects
}
\examples{
###
}
