% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mulitpleAlignment.R
\name{.multipleAlignment}
\alias{.multipleAlignment}
\title{Multiple Alignment of \code{MicrobiomeFunction} Objects}
\usage{
.multipleAlignment(coms)
}
\arguments{
\item{coms}{Named list of MiCo objects.}
}
\value{
A list containing the alignment data.
}
\description{
Multiple Alignment of \code{MicrobiomeFunction} Objects
}
