% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseAlignment.R
\name{.pairwiseAlignment}
\alias{.pairwiseAlignment}
\title{Pairwise Alignment of \code{MicrobiomeFunction} Objects}
\usage{
.pairwiseAlignment(coms)
}
\arguments{
\item{coms}{Named list of MiCo objects.}
}
\value{
List of pairwise alignments.
}
\description{
\itemize{
\item This method only allows for the alignment of edges which are present in both communities.
\item 
}
}
\examples{
###
}
