% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareAlignments.R
\name{.rxnsPerLevel}
\alias{.rxnsPerLevel}
\title{Get Count of Reactions Per Level of Alignment}
\usage{
.rxnsPerLevel(object, name)
}
\arguments{
\item{object}{MiCoAl object}

\item{name}{Name of the object}
}
\value{
A tibble with levels and count of reaction
}
\description{
Get Count of Reactions Per Level of Alignment
}
