% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreAlignment.R
\name{.scoreAlignment}
\alias{.scoreAlignment}
\title{Scoring Algorithm for MiCoAl Objects}
\usage{
.scoreAlignment(a)
}
\arguments{
\item{a}{alignment list}
}
\value{
Tibble containing scores for every level of alignment
}
\description{
Takes an alignment object and calculates the overall alignment score.
}
