% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exCo1}
\alias{exCo1}
\title{Example Community 1}
\format{
A tibble with the columns:
\describe{
\item{MO}{Microorganism},
\item{met}{Metabolite}
\item{flux}{Flux of the metabolite}}
}
\source{
\url{https://github.com/sirno/misosoup}
}
\usage{
exCo1
}
\description{
Example community taken from \code{misosoup}.
}
\examples{
exCo1

}
\keyword{datasets}
