% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findEdges.R
\name{findEdges}
\alias{findEdges}
\title{Get All Edges of a Community}
\usage{
findEdges(tb)
}
\arguments{
\item{tb}{Tibble describing a microbial community}
}
\value{
A hash containing the edges of the community. Returns a list of
two hashs, one for edges between metabolites and one for species.
}
\description{
Takes a tibble describing a microbial community as input and returns a
list with two hashs, one giving the edges between metabolites and the other
between species.
}
\examples{
findEdges(ac_A1R12_1)
}
