% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsmc-package.R
\docType{package}
\name{fsmc-package}
\alias{fsmc}
\alias{fsmc-package}
\title{fsmc: Functional Similarity of Microbial Communities}
\description{
Analysis of the functional (dis-) similarity of microbial communities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://admarhi.github.io/fsmc/}
}

}
\author{
\strong{Maintainer}: Adrian Hirt \email{am.hirt@pm.me} (\href{https://orcid.org/0009-0008-1929-6629}{ORCID})

}
\keyword{internal}
