% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edges.R
\name{get_edges}
\alias{get_edges}
\title{Get All Edges of a Community}
\usage{
get_edges(tb, silent = FALSE)
}
\arguments{
\item{tb}{Tibble describing a microbial community}

\item{silent}{Logical to toggle output}
}
\value{
A list of two hashs, one for edges between metabolites and one for
species.
}
\description{
Takes a tibble describing a microbial community as input and returns a
list with two hashs, one giving the edges between metabolites and the other
between species.
}
\examples{
get_edges(ac_A1R12_1)
}
