% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neighbs.R
\name{get_neighbs}
\alias{get_neighbs}
\title{Find All Neighbors in a Community}
\usage{
get_neighbs(df, silent = FALSE)
}
\arguments{
\item{df}{Data frame describing a microbial community}

\item{silent}{Logical to toggle output}
}
\value{
A list with neighbor data
}
\description{
Takes a data frame describing a microbial community as input and returns a
list with both neighbors and ###
}
\examples{

get_neighbs(exCo1)
}
