% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-getCo.R
\name{getCo}
\alias{getCo}
\alias{getCo,MiCo-method}
\alias{getCo,MiCoAl-method}
\title{Get the Community}
\usage{
getCo(object)

\S4method{getCo}{MiCo}(object)

\S4method{getCo}{MiCoAl}(object)
}
\arguments{
\item{object}{An object of class MiCoAl}
}
\value{
A tibble with the community data.

A list with the community data.
}
\description{
Returns the community of a single MiCo object in a tibble format or a list
of communities in tibble format for MiCoAl objects.
}
\section{Functions}{
\itemize{
\item \code{getCo(MiCo)}: Get the Community

\item \code{getCo(MiCoAl)}: Get the Community

}}
\examples{
c1 <- newMiCo(ac_A1R12_1)
c2 <- newMiCo(ac_A1R12_2)
a <- newMiCoAl(c1, c2)

getCo(c1)
getCo(a)
}
