% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-getEdges.R
\name{getEdges}
\alias{getEdges}
\alias{getEdges,MiCo-method}
\title{Get Edges From a MiCo (Microbial Community) Object}
\usage{
getEdges(object, unique = FALSE)

\S4method{getEdges}{MiCo}(object, unique = FALSE)
}
\arguments{
\item{object}{An object of class MiCo}
}
\value{
A list of edges in the community.
}
\description{
Get Edges From a MiCo (Microbial Community) Object
}
\section{Functions}{
\itemize{
\item \code{getEdges(MiCo)}: Get Edges From a MiCo (Microbial Community) Object

}}
