% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getFlux}
\alias{getFlux}
\alias{getFlux,MicrobialCommunity-method}
\title{Get flux from MicrobialCommunity object}
\usage{
getFlux(object)

\S4method{getFlux}{MicrobialCommunity}(object)
}
\arguments{
\item{object}{An object of class MicrobialCommunity.}
}
\value{
A numeric vector representing the fluxes.

A numeric vector representing the fluxes.
}
\description{
Get flux from MicrobialCommunity object
}
\section{Functions}{
\itemize{
\item \code{getFlux(MicrobialCommunity)}: Get flux from MicrobialCommunity object

}}
