% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getMet}
\alias{getMet}
\alias{getMet,MicrobialCommunity-method}
\title{Get met from MicrobialCommunity object}
\usage{
getMet(object)

\S4method{getMet}{MicrobialCommunity}(object)
}
\arguments{
\item{object}{An object of class MicrobialCommunity.}
}
\value{
A character vector representing the metabolites.

A character vector representing the metabolites.
}
\description{
Get met from MicrobialCommunity object
}
\section{Functions}{
\itemize{
\item \code{getMet(MicrobialCommunity)}: Get met from MicrobialCommunity object

}}
