% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-getMO.R
\name{getMO}
\alias{getMO}
\alias{getMO,MiCo-method}
\title{Get MO from c object}
\usage{
getMO(object, unique = FALSE)

\S4method{getMO}{MiCo}(object, unique = FALSE)
}
\arguments{
\item{object}{An object of class MiCo}

\item{unique}{Logical to toggle unique output.}
}
\value{
A character vector representing the microorganisms.

A character vector representing the microorganisms.
}
\description{
Get MO from c object
}
\section{Functions}{
\itemize{
\item \code{getMO(MiCo)}: Get MO from MicrobialCommunity object

}}
