% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-getSpecies.R
\name{getSpecies}
\alias{getSpecies}
\alias{getSpecies,MiCo-method}
\alias{getSpecies,MiCoAl-method}
\title{Get Species From a MiCo (Microbial Community) Object}
\usage{
getSpecies(object)

\S4method{getSpecies}{MiCo}(object)

\S4method{getSpecies}{MiCoAl}(object)
}
\arguments{
\item{object}{An object of class MiCoAl}
}
\value{
A character vector representing the microorganisms.

A character vector representing the microorganisms.
}
\description{
Get Species From a MiCo (Microbial Community) Object
}
\section{Functions}{
\itemize{
\item \code{getSpecies(MiCo)}: Get Species From a MiCo (Microbial Community) Object

\item \code{getSpecies(MiCoAl)}: Get Species From a MiCo (Microbial Community) Object

}}
