% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_alignment.R
\name{MiCo.align}
\alias{MiCo.align}
\title{Functional Similarity of Microbial Communities}
\usage{
MiCo.align(co_list, edges_n = TRUE, edges_js = TRUE, flux_diff = TRUE)
}
\arguments{
\item{co_list}{List containing communities}

\item{edges_n}{Toggle for number of edges output}

\item{edges_js}{Toggle Jaccard Score Calculation}

\item{flux_diff}{Get the difference of fluxes}
}
\value{
List with alignment values
}
\description{
Functional Similarity of Microbial Communities
}
\examples{
#
}
