% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_alignment.R
\name{MiCo.align.v2}
\alias{MiCo.align.v2}
\title{Functional Similarity of Microbial Communities (v2)}
\usage{
MiCo.align.v2(co_list)
}
\arguments{
\item{co_list}{List containing communities}
}
\value{
List with alignment values
}
\description{
Functional Similarity of Microbial Communities (v2)
}
\examples{
#
}
