% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-MiCo.R
\name{MiCo}
\alias{MiCo}
\title{Constructor Function for MiCo (Microbial Community) Objects}
\usage{
MiCo(
  path = NULL,
  name = NULL,
  species = character(),
  metabolites = character(),
  fluxes = numeric()
)
}
\arguments{
\item{path}{Path to a csv file storing community data. Not required when
using MO, met, flux to specify the data.}

\item{name}{A character vector representing the name of the community.}

\item{species}{character vector representing the microorganisms present in the
community.}

\item{metabolites}{A character vectors representing the metabolites consumed and
produced within the community.}

\item{fluxes}{A numeric vector representing the fluxes of each metabolite in
the community.}
}
\description{
Constructor function for MiCo objects. This function can be used
to create a new MiCo object from a CSV file or from vectors of
microorganisms, metabolites and fluxes.
}
