% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-mico_align.R
\name{mico_align}
\alias{mico_align}
\title{Functional Alignment of Microbial Communities}
\usage{
mico_align(..., pairwise = FALSE)
}
\arguments{
\item{...}{Microbial communities}

\item{pairwise}{Logical indicating whether to perform pairwise alignment.}
}
\value{
List with alignment values
}
\description{
Functional Alignment of Microbial Communities
}
