% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-MiCoAl.R
\docType{class}
\name{MiCoAl-class}
\alias{MiCoAl-class}
\title{Microbial Community Alignment Class}
\description{
A class to represent an alignment of multiple microbial
communities
}
\section{Slots}{

\describe{
\item{\code{communities}}{A list of objects of type MicrobialCommunity.}

\item{\code{alignment}}{A list containing the alignment data.}
}}

