% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-MiCoAl.R
\name{MiCoAl}
\alias{MiCoAl}
\title{Constructor for Microbial Community Alignment Objects}
\usage{
MiCoAl(..., pairwise = FALSE)
}
\arguments{
\item{...}{MiCo objects to be aligned.}

\item{pairwise}{Logical indicating whether to perform pairwise alignment.}
}
\description{
Constructor for Microbial Community Alignment Objects
}
