% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{MicrobialCommunityAlignment-class}
\alias{MicrobialCommunityAlignment-class}
\title{MicrobialCommunityAlignment-class}
\description{
A class to represent an alignment of multiple microbial
communities
}
\section{Slots}{

\describe{
\item{\code{mico}}{A list of objects of type MicrobialCommunity.}

\item{\code{alignment_info}}{A character vector representing alignment-related
information.}
}}

