% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{MicrobialCommunityAlignment}
\alias{MicrobialCommunityAlignment}
\title{Constructor Function for MicrobialCommunityAlignment Objects}
\usage{
MicrobialCommunityAlignment(mico, alignment_info)
}
\arguments{
\item{mico}{A list of of objects of type MicrobialCommunity.}

\item{alignment_info}{A character vector giving additional information for
the alignment.}
}
\description{
Constructor Function for MicrobialCommunityAlignment Objects
}
