% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeFunctionSet.R
\name{MicrobiomeFunctionSet}
\alias{MicrobiomeFunctionSet}
\title{Set of \code{MicrobiomeFunction} Objects}
\usage{
MicrobiomeFunctionSet(..., name = NA_character_)
}
\arguments{
\item{...}{Lists or individual \code{MicrobiomeFunction} objects.}

\item{name}{Character scalar giving the name of the set.}
}
\description{
Set of \code{MicrobiomeFunction} Objects
}
