% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{misosoup}
\alias{misosoup}
\title{Example Data from MiSoSoup}
\format{
A tibble with the columns:
\describe{
\item{species}{Microorganisms},
\item{metabolites}{Metabolites}
\item{fluxes}{Fluxes of the metabolite}
}
}
\usage{
misosoup
}
\description{
List of 56 solutions from MiSoSoup for different focal strains in different
media.
}
\keyword{datasets}
