% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-multiple_alignment.R
\name{multiple_alignment}
\alias{multiple_alignment}
\title{Multiple Alignment of MiCo (Microbial Community) Objects}
\usage{
multiple_alignment(coms)
}
\arguments{
\item{coms}{Named list of MiCo objects.}
}
\value{
A list containing the alignment data.
}
\description{
Multiple Alignment of MiCo (Microbial Community) Objects
}
\examples{
###
}
