% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newMiCoAl.R
\name{newMiCoAl}
\alias{newMiCoAl}
\title{Constructor for Microbial Community Alignment Objects}
\usage{
newMiCoAl(..., pairwise = FALSE, comment = NULL)
}
\arguments{
\item{...}{MiCo objects to be aligned.}

\item{pairwise}{Logical indicating whether to perform pairwise alignment.}

\item{comment}{A comment to be added to the alignment.}
}
\description{
Constructor for Microbial Community Alignment Objects
}
