% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_align.R
\name{pairwise_align}
\alias{pairwise_align}
\title{Pairwise Alignment of Microbial Communities}
\usage{
pairwise_align(mi_co_list)
}
\arguments{
\item{mi_co_list}{A list of microbial community objects}
}
\value{
A list of information about the alignment of the communities.
}
\description{
Pairwise Alignment of Microbial Communities
}
\examples{
### Todo
}
