% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-pairwise_alignment.R
\name{pairwise_alignment}
\alias{pairwise_alignment}
\title{Pairwise Alignment of MiCo (Microbial Community) Objects}
\usage{
pairwise_alignment(coms)
}
\arguments{
\item{coms}{Named list of MiCo objects.}
}
\value{
List of pairwise alignments.
}
\description{
Pairwise Alignment of MiCo (Microbial Community) Objects
}
\examples{
###
}
