% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotMF.R
\name{pivotMF}
\alias{pivotMF}
\title{Functions to Change the orientation of MiCo Input Data}
\usage{
pivotMF(tb, species, from, to, flux)
}
\arguments{
\item{tb}{Tibble with data on a microbial community in
long or short format to be}

\item{species}{Column name of the species column}

\item{from}{the name of the column specifying the met consumed}

\item{to}{the name of the column specifying the met excreted}

\item{flux}{Column name of the flux column}
}
\description{
Wrapper function around tidyr's \code{pivot_longer()} function to facilitate
the easy transformation into the correct data format for MiCo construction.
}
\examples{
tb <- tibble::tribble(
  ~uptake, ~secretion, ~flux, ~species,
  "m1", "m2", 1, "s1",
  "m2", "m3", 1, "s2",
  "m3", "m4", 1, "s3",
  "m4", "m1", 1, "s4"
)

pivotMF(
  tb = tb,
  species = "species",
  from = "uptake",
  to = "secretion",
  flux = "flux"
)
}
