% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotMiCo.R
\name{pivotMiCo}
\alias{pivotMiCo}
\title{Functions to Change the orientation of MiCo Input Data}
\usage{
pivotMiCo(tb, species, from, to, flux)
}
\arguments{
\item{tb}{Tibble with data on a microbial community in
long or short format to be}

\item{species}{Column name of the species column}

\item{from}{the name of the column specifying the met consumed}

\item{to}{the name of the column specifying the met excreted}

\item{flux}{Column name of the flux column}
}
\description{
Wrapper function around tidyr's \code{pivot_longer()} function to facilitate
the easy transformation into the correct data format for MiCo construction.
}
