% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignmentHeatmap.R
\name{plotAlignmentHeatmap}
\alias{plotAlignmentHeatmap}
\title{MiCoAl Heatmap}
\usage{
plotAlignmentHeatmap(object, frac)
}
\arguments{
\item{object}{An object of class MiCoAl.}

\item{frac}{Minimum relative weight of maximum weight to plot.}
}
\value{
A ggplot heatmap
}
\description{
MiCoAl Heatmap
}
