% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignmentNetwork.R
\name{plotAlignmentNetwork}
\alias{plotAlignmentNetwork}
\title{MiCoAl Network Plot}
\usage{
plotAlignmentNetwork(object, frac)
}
\arguments{
\item{object}{An object of class MiCoAl.}

\item{frac}{Minimum relative weight of maximum weight to plot.}
}
\value{
An igraph network
}
\description{
MiCoAl Network Plot
}
