% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-syn_data.R
\name{rand_mo_data}
\alias{rand_mo_data}
\title{Generate Random Synthetic Communities}
\usage{
rand_mo_data(
  n_mo,
  max_met,
  n_co = 2,
  scale_fac = 2,
  seed = 123,
  dead_ends = FALSE
)
}
\arguments{
\item{n_mo}{Number of micro organisms}

\item{max_met}{Maximum number of metabolites in the communities}

\item{n_co}{Number of communities}

\item{scale_fac}{Scaling factor}

\item{seed}{Seed for reproducibility}

\item{dead_ends}{Logical value to toggle dead ends in data}
}
\value{
List with \code{n_co} number of communities.
}
\description{
Generate Random Synthetic Communities
}
\examples{
#
}
