% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-show.R
\name{show,MiCo-method}
\alias{show,MiCo-method}
\title{Show method for Microbial Community objects}
\usage{
\S4method{show}{MiCo}(object)
}
\arguments{
\item{object}{An object of class MiCo}
}
\description{
Show method for Microbial Community objects
}
