% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-show.R
\name{show,MiCoAl-method}
\alias{show,MiCoAl-method}
\title{Show method for Microbial Community Alignment Objects}
\usage{
\S4method{show}{MiCoAl}(object)
}
\arguments{
\item{object}{An object of class MiCoAl}
}
\description{
Show method for Microbial Community Alignment Objects
}
