% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,MicrobialCommunity-method}
\alias{show,MicrobialCommunity-method}
\title{Show method for MicrobialCommunity objects}
\usage{
\S4method{show}{MicrobialCommunity}(object)
}
\arguments{
\item{object}{An object of class MicrobialCommunity.}
}
\description{
Show method for MicrobialCommunity objects
}
