% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-show.R
\name{show,MicrobiomeFunction-method}
\alias{show,MicrobiomeFunction-method}
\title{Show Method for \code{MicrobiomeFunction} Object}
\usage{
\S4method{show}{MicrobiomeFunction}(object)
}
\arguments{
\item{object}{An object of class \code{MicrobiomeFunction}}
}
\description{
Show Method for \code{MicrobiomeFunction} Object
}
