% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-show.R
\name{show,MicrobiomeFunctionAlignment-method}
\alias{show,MicrobiomeFunctionAlignment-method}
\title{Show method for \code{MicrobiomeFunctionAlignment} Objects}
\usage{
\S4method{show}{MicrobiomeFunctionAlignment}(object)
}
\arguments{
\item{object}{a \code{MicrobiomeFunctionAlignment} object.}
}
\description{
Show method for \code{MicrobiomeFunctionAlignment} Objects
}
