% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-syn_community.R
\name{syn_community}
\alias{syn_community}
\title{Generate Random Synthetic Microbial Communities}
\usage{
syn_community(
  name,
  n_species,
  max_met,
  scale_fac = 2,
  seed = FALSE,
  dead_ends = FALSE,
  MiCo = TRUE
)
}
\arguments{
\item{name}{Character string giving the desired name of the community.}

\item{n_species}{Number of species in the community}

\item{max_met}{Maximum number of metabolites in the communities}

\item{scale_fac}{Scaling factor}

\item{seed}{Seed for reproducibility}

\item{dead_ends}{Logical value to toggle dead ends in data}

\item{MiCo}{Logical value to toggle return of MiCo object or tibble.}
}
\value{
List with \code{n_co} number of communities.
}
\description{
A function that creates synthetic data suitable for demonstration purposes of the \code{fsmc} package.
}
\examples{
syn_community("Ex. Community", n_species = 5, max_met = 10)
}
