% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-alignmentHeatmap.R
\name{alignmentHeatmap}
\alias{alignmentHeatmap}
\alias{alignmentHeatmap,MiCoAl-method}
\title{MiCoAl Heatmap}
\usage{
alignmentHeatmap(object, frac)

\S4method{alignmentHeatmap}{MiCoAl}(object, frac)
}
\arguments{
\item{object}{An object of class MiCoAl.}

\item{frac}{Minimum relative weight of maximum weight to plot.}
}
\value{
A ggplot heatmap
}
\description{
MiCoAl Heatmap
}
\examples{
###
}
