% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreAlignment.R
\name{.scoreAlignment}
\alias{.scoreAlignment}
\title{Scoring Algorithm for MiCoAl Objects}
\usage{
.scoreAlignment(a, min_rxn, min_mb)
}
\arguments{
\item{a}{alignment list}

\item{min_rxn}{Numeric scalar giving the minimum fraction of reactions}

\item{min_mb}{Numeric scalar giving the minimum fraction of microbiomes}
}
\value{
Tibble containing scores for every level of alignment
}
\description{
Takes an alignment object and calculates the overall alignment score.
}
