% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-getCo.R
\name{getCo}
\alias{getCo}
\alias{getCo,MiCo-method}
\title{Get the Community}
\usage{
getCo(object)

\S4method{getCo}{MiCo}(object)
}
\arguments{
\item{object}{An object of class MiCo}
}
\value{
A tibble with the community data.

A tibble with the community data.
}
\description{
Get the Community
}
\section{Functions}{
\itemize{
\item \code{getCo(MiCo)}: Get the Community

}}
\examples{
### Write examples for both classes
### Write examples for both classes
}
