% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-getCo.R
\name{getCo}
\alias{getCo}
\alias{getCo,MicrobiomeFunction-method}
\title{Get the Community}
\usage{
getCo(object)

\S4method{getCo}{MicrobiomeFunction}(object)
}
\arguments{
\item{object}{An object of class \code{MicrobiomeFunction}}
}
\value{
A list with the community data.
}
\description{
Returns the community of a single \code{MicrobiomeFunction} object in a \
tibble format or a list of communities in tibble format for
\code{MicrobiomeFunctionAlignment} objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{getCo(MicrobiomeFunction)}: Get the Community

}}
\examples{
c1 <- newMiCo(ac_A1R12_1)
c2 <- newMiCo(ac_A1R12_2)
a <- newMiCoAl(c1, c2)

getCo(c1)
getCo(a)
}
