% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-getEdges.R
\name{getEdges}
\alias{getEdges}
\alias{getEdges,MicrobiomeFunction-method}
\title{Get Edges From a \code{MicrobiomeFunction} Object}
\usage{
getEdges(object)

\S4method{getEdges}{MicrobiomeFunction}(object)
}
\arguments{
\item{object}{a \code{MicrobiomeFunction} object}
}
\value{
A list of edges in the community.
}
\description{
Get Edges From a \code{MicrobiomeFunction} Object
}
\section{Methods (by class)}{
\itemize{
\item \code{getEdges(MicrobiomeFunction)}: Get Edges From a \code{MicrobiomeFunction} Object

}}
