% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-getEdges.R
\name{getEdges}
\alias{getEdges}
\alias{getEdges,MiCo-method}
\title{Get Edges From a MiCo (Microbial Community) Object}
\usage{
getEdges(object, type = NULL)

\S4method{getEdges}{MiCo}(object, type = NULL)
}
\arguments{
\item{object}{An object of class MiCo}

\item{type}{Character string indicating which type of edges to return. Can
be either of \code{c("species", "metabolites")}, and defaults to \code{NULL.}
If \code{NULL}, both types of edges are returned.}
}
\value{
A list of edges in the community.
}
\description{
Get Edges From a MiCo (Microbial Community) Object
}
\section{Functions}{
\itemize{
\item \code{getEdges(MiCo)}: Get Edges From a MiCo (Microbial Community) Object

}}
