% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-getFlux.R
\name{getFlux}
\alias{getFlux}
\alias{getFlux,MiCo-method}
\title{Get Fluxes from MiCo Object}
\usage{
getFlux(object)

\S4method{getFlux}{MiCo}(object)
}
\arguments{
\item{object}{An object of class MicrobialCommunity.}
}
\value{
A numeric vector representing the fluxes.

A numeric vector representing the fluxes.
}
\description{
Get Fluxes from MiCo Object
}
\section{Functions}{
\itemize{
\item \code{getFlux(MiCo)}: Get flux from MicrobialCommunity object

}}
