% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-getMet.R
\name{getMet}
\alias{getMet}
\alias{getMet,MiCo-method}
\title{Get Metabolites from MiCo object}
\usage{
getMet(object)

\S4method{getMet}{MiCo}(object)
}
\arguments{
\item{object}{An object of class MicrobialCommunity.}
}
\value{
A character vector representing the metabolites.

A character vector representing the metabolites.
}
\description{
Get Metabolites from MiCo object
}
\section{Functions}{
\itemize{
\item \code{getMet(MiCo)}: Get met from MicrobialCommunity object

}}
