% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-getMet.R
\name{getMet}
\alias{getMet}
\alias{getMet,MicrobiomeFunction-method}
\alias{getMet,MicrobiomeFunctionAlignment-method}
\title{Get Metabolites}
\usage{
getMet(object)

\S4method{getMet}{MicrobiomeFunction}(object)

\S4method{getMet}{MicrobiomeFunctionAlignment}(object)
}
\arguments{
\item{object}{a \code{MicrobiomeFunction} or
\code{MicrobiomeFunctionAlignment} object}
}
\value{
A character vector representing the metabolites.
}
\description{
Get Metabolites
}
