% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-getMet.R
\name{getMet}
\alias{getMet}
\alias{getMet,MiCo-method}
\alias{getMet,MiCoAl-method}
\title{Get Metabolites}
\usage{
getMet(object, unique = TRUE)

\S4method{getMet}{MiCo}(object, unique = TRUE)

\S4method{getMet}{MiCoAl}(object)
}
\arguments{
\item{object}{An object of class MiCo or MiCoAl}

\item{unique}{Logical to toggle unique output.}
}
\value{
A character vector representing the metabolites.
}
\description{
Get Metabolites
}
