% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-getMet.R
\name{getMet}
\alias{getMet}
\alias{getMet,MiCo-method}
\title{Get Metabolites from MiCo object}
\usage{
getMet(object, unique = FALSE)

\S4method{getMet}{MiCo}(object, unique = FALSE)
}
\arguments{
\item{object}{An object of class MiCo}

\item{unique}{Logical to toggle unique output.}
}
\value{
A character vector representing the metabolites.
}
\description{
Get Metabolites from MiCo object
}
