% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getMO}
\alias{getMO}
\alias{getMO,MicrobialCommunity-method}
\title{Get MO from MicrobialCommunity object}
\usage{
getMO(object, unique = FALSE)

\S4method{getMO}{MicrobialCommunity}(object, unique = FALSE)
}
\arguments{
\item{object}{An object of class MicrobialCommunity.}

\item{unique}{Liogical to toggle unique output.}

\item{MicrobialCommunity}{}
}
\value{
A character vector representing the microorganisms.

A character vector representing the microorganisms.
}
\description{
Get MO from MicrobialCommunity object
}
\section{Functions}{
\itemize{
\item \code{getMO(MicrobialCommunity)}: Get MO from MicrobialCommunity object

}}
