% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-getSpecies.R
\name{getSpecies}
\alias{getSpecies}
\alias{getSpecies,MicrobiomeFunction-method}
\alias{getSpecies,MicrobiomeFunctionAlignment-method}
\title{Return Species in a Microbiome}
\usage{
getSpecies(object)

\S4method{getSpecies}{MicrobiomeFunction}(object)

\S4method{getSpecies}{MicrobiomeFunctionAlignment}(object)
}
\arguments{
\item{object}{a \code{MicrobiomeFunctionAlignment} Object}
}
\value{
A character vector representing the microorganisms.

A character vector representing the microorganisms.
}
\description{
Return Species in a Microbiome
}
\section{Functions}{
\itemize{
\item \code{getSpecies(MicrobiomeFunction)}: Return Species in a Microbiome

\item \code{getSpecies(MicrobiomeFunctionAlignment)}: Return Species in a Microbiome

}}
