% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-getMO.R
\name{getSpecies}
\alias{getSpecies}
\alias{getSpecies,MiCo-method}
\title{Get Species From a MiCo (Microbial Community) Object}
\usage{
getSpecies(object, unique = FALSE)

\S4method{getSpecies}{MiCo}(object, unique = FALSE)
}
\arguments{
\item{object}{An object of class MiCo}

\item{unique}{Logical to toggle unique output.}
}
\value{
A character vector representing the microorganisms.

A character vector representing the microorganisms.
}
\description{
Get Species From a MiCo (Microbial Community) Object
}
\section{Functions}{
\itemize{
\item \code{getSpecies(MiCo)}: Get Species From a MiCo (Microbial Community) Object

}}
