% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-MiCo.R
\docType{class}
\name{MiCo-class}
\alias{MiCo-class}
\title{MicrobialCommunity-class}
\description{
A class to represent communities of microorganisms.
}
\section{Slots}{

\describe{
\item{\code{MO}}{A character vector representing the microorganisms present in the
community.}

\item{\code{met}}{A character vectors representing the metabolites consumed and
produced within the community.}

\item{\code{flux}}{A numeric vector representing the fluxes of each metabolite in
the community.}
}}

