% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-MiCo.R
\name{MiCo}
\alias{MiCo}
\title{Constructor Function for MicrobialCommunity Objects}
\usage{
MiCo(path = NULL, MO = character(), met = character(), flux = numeric())
}
\arguments{
\item{path}{Path to a csv file storing community data. Not required when
using MO, met, flux to specify the data.}

\item{MO}{character vector representing the microorganisms present in the
community.}

\item{met}{A character vectors representing the metabolites consumed and
produced within the community.}

\item{flux}{A numeric vector representing the fluxes of each metabolite in
the community.}
}
\description{
Constructor Function for MicrobialCommunity Objects
}
