% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-MiCoAl.R
\name{MiCoAl}
\alias{MiCoAl}
\title{Constructor for Microbial Community Alignment Objects}
\usage{
MiCoAl(communities)
}
\arguments{
\item{communities}{A list of of objects of type MiCo}
}
\description{
Constructor for Microbial Community Alignment Objects
}
