% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{MicrobialCommunityAlignment-class}
\alias{MicrobialCommunityAlignment-class}
\title{MicrobialCommunityAlignment-class}
\description{
A class to represent an alignment of multiple microbial
communities
}
\section{Slots}{

\describe{
\item{\code{mi_co}}{A list of objects of type MicrobialCommunity.}

\item{\code{output_data}}{A list containing the alignment data.}
}}

