% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{MicrobialCommunityAlignment}
\alias{MicrobialCommunityAlignment}
\title{Constructor Function for MicrobialCommunityAlignment Objects}
\usage{
MicrobialCommunityAlignment(mi_co)
}
\arguments{
\item{mi_co}{A list of of objects of type MicrobialCommunity.
the alignment.}
}
\description{
Constructor Function for MicrobialCommunityAlignment Objects
}
